Major cleanup of Sections & Views. Observers properly hooked up to appropriate states.
